Add pages

+++ src/components/post-form/PostForm.jsx
@@ -3,6 +3,7 @@ import { useSelector } from "react-redux";
 import { useForm } from "react-hook-form";
 import appwriteService from "../../appwrite/config";
 import { useNavigate } from "react-router-dom";
+import React, { useCallback } from "react";
 
 export default function PostForm({ post }) {
   const { register, handleSubmit, watch, control, setValue, getValues } =
@@ -19,35 +20,45 @@ export default function PostForm({ post }) {
   const userData = useSelector((state) => state.auth.userData);
 
   const submit = async (data) => {
-    if (post) {
-      const file = data.image[0]
-        ? await appwriteService.uploadFile(data.image[0])
-        : null;
-
-      if (file) {
-        await appwriteService.deleteFile(post.featuredImage);
+    try {
+      if (!userData) {
+        console.error("User not logged in");
+        return;
       }
 
-      const dbPost = await appwriteService.updatePost(post.$id, {
-        ...data,
-        featuredImage: file ? file.$id : undefined,
-      });
+      if (post) {
+        const file = data.image[0]
+          ? await appwriteService.uploadFile(data.image[0])
+          : null;
 
-      if (dbPost) {
-        navigate(`/post/${dbPost.$id}`);
-      }
-    } else {
-      const file = await appwriteService.uploadFile(data.image[0]);
+        if (file) {
+          await appwriteService.deleteFile(post.featuredImage);
+        }
 
-      if (file) {
-        data.featuredImage = file.$id;
-        const dbPost = await appwriteService.createPost({
+        const dbPost = await appwriteService.updatePost(post.$id, {
           ...data,
-          userId: userData.$id,
+          featuredImage: file ? file.$id : undefined,
         });
 
-        if (dbPost) navigate(`/post/${dbPost.$id}`);
+        if (dbPost) {
+          navigate(`/post/${dbPost.$id}`);
+        }
+      } else {
+        const file = await appwriteService.uploadFile(data.image[0]);
+
+        if (file) {
+          data.featuredImage = file.$id;
+          
+          const dbPost = await appwriteService.createPost({
+            ...data,
+            userId: userData.$id,
+          });
+
+          if (dbPost) navigate(`/post/${dbPost.$id}`);
+        }
       }
+    } catch (error) {
+      console.error("POST SUBMIT ERROR, ", error);
     }
   };
